# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Auth Service.",
        "title": "Auth Service API3",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/twits/addLike": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Foydalanuvchi tomonidan tanlangan twitga like qo'shadi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Likes"
                ],
                "summary": "Twitga like qo'shish",
                "parameters": [
                    {
                        "description": "Twit ID ma'lumotlari",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/twit.AddLikeReq1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like qo'shish muvaffaqiyatli bo'ldi",
                        "schema": {
                            "$ref": "#/definitions/twit.AddLikeResp"
                        }
                    },
                    "400": {
                        "description": "Ma'lumotlarni olishda xatolik",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "AddLike request error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/twits/createTwit": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Foydalanuvchi tomonidan yangi twit yaratadi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Twits"
                ],
                "summary": "Yangi twit yaratish",
                "parameters": [
                    {
                        "description": "Twit ma'lumotlari",
                        "name": "CreateTwitReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/twit.CreateTwitReq1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/twit.CreateTwitResp"
                        }
                    },
                    "400": {
                        "description": "Ma'lumotlarni olishda xatolik",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "CreateTwit request error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/twits/deleteTwit/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Foydalanuvchi tomonidan yaratilgan twitni o'chiradi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Twits"
                ],
                "summary": "Twitni o'chirish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Twit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/twit.Message"
                        }
                    },
                    "500": {
                        "description": "DeleteTwit request error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/twits/getFollowerTwit": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Foydalanuvchining kuzatuvchilari tomonidan yaratilgan barcha twitlarni qaytaradi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Twits"
                ],
                "summary": "Kuzatuvchilar twitlarini olish",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/twit.GetTwitsResp"
                        }
                    },
                    "500": {
                        "description": "GetFollowerTwit request error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/twits/getUserTwit": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Foydalanuvchining barcha twitlarini qaytaradi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Twits"
                ],
                "summary": "Foydalanuvchi twitlarini olish",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/twit.GetTwitsResp"
                        }
                    },
                    "500": {
                        "description": "GetTwits request error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/twits/unlike": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Foydalanuvchi tomonidan tanlangan twitdan like olib tashlanadi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Likes"
                ],
                "summary": "Twitdan like olib tashlash",
                "parameters": [
                    {
                        "description": "Twit ID ma'lumotlari",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/twit.DeleteTwitReq1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli olib tashlandi",
                        "schema": {
                            "$ref": "#/definitions/twit.Message"
                        }
                    },
                    "400": {
                        "description": "Ma'lumotlarni olishda xatolik",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "RemoveLike request error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/twits/updateTwit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Foydalanuvchi tomonidan mavjud twitni yangilaydi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Twits"
                ],
                "summary": "Mavjud twitni yangilash",
                "parameters": [
                    {
                        "description": "Yangilash uchun twit ma'lumotlari",
                        "name": "UpdateTwitReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/twit.UpdateReq1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/twit.UpdateTwitResp"
                        }
                    },
                    "400": {
                        "description": "Ma'lumotlarni olishda xatolik",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "UpdateTwit request error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "twit.AddLikeReq1": {
            "type": "object",
            "properties": {
                "twit_id": {
                    "type": "string"
                }
            }
        },
        "twit.AddLikeResp": {
            "type": "object",
            "properties": {
                "massage": {
                    "type": "string"
                }
            }
        },
        "twit.CreateTwitReq1": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "media": {
                    "type": "string"
                }
            }
        },
        "twit.CreateTwitResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "twit.DeleteTwitReq1": {
            "type": "object",
            "properties": {
                "twit_id": {
                    "type": "string"
                }
            }
        },
        "twit.GetTwitsResp": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "twits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/twit.Twit"
                    }
                }
            }
        },
        "twit.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "twit.Twit": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "twit.UpdateReq1": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media": {
                    "type": "string"
                }
            }
        },
        "twit.UpdateTwitResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}